Modify batches to build and test flowgen utility.

+++ build.sh
@@ -3,6 +3,7 @@
 rm $GOPATH/bin/master
 rm $GOPATH/bin/servant
 rm $GOPATH/bin/show
+rm $GOPATH/bin/flowgen
 
 # *****************
 # Unit test
@@ -38,15 +39,24 @@ if [ "$?" -ne "0" ] ; then
     exit 1
 fi
 
+echo "show utility building..."
+go install cuto/flowgen
+if [ "$?" -ne "0" ] ; then
+    echo "flowgen build NG."
+    exit 1
+fi
+
 chmod a+x $GOPATH/bin/*
 
 rm $GOPATH/cutoroot/bin/master
 rm $GOPATH/cutoroot/bin/servant
 rm $GOPATH/cutoroot/bin/show
+rm $GOPATH/cutoroot/bin/flowgen
 
 cp $GOPATH/bin/master $GOPATH/cutoroot/bin
 cp $GOPATH/bin/servant $GOPATH/cutoroot/bin
 cp $GOPATH/bin/show $GOPATH/cutoroot/bin
+cp $GOPATH/bin/flowgen $GOPATH/cutoroot/bin
 
 exit 0
